Lift parallax helpers into p9-core::coords
